requirements.txt
aiogram~=2.12
aiohttp

python-dotenv~=0.15.0
